Concord: DAML PoC integration
This is the final change to bring the DAML PoC onto Concord.
Take a look at Concord's startup, we use a different API service for DAML (gRPC
with its services defined in `concord/src/daml/grpc_services`) and a different
commands handler (`concord/src/daml/cmd_handler`) depending on the
configuration flag `daml_enable`. With the later we needed to define a
for `Blockchain::ICommandsHanlder` in order to free the memory of either
version on program exit.

How to test? Use the newly provided `docker-compose-daml.yml` to start up the
deployment. It requires to build daml's ledger api image from gitlab first.
Step-by-step instructions are in the `yml` file. Also, make sure you set
`daml_enable: true` in Concord's config files. Once deployed, log in to the
ledger api service and run the semantic test runner as described in the README.