Add additional stability fix and config flexibility
This patch adds 2 misc changes to provisioning service.

1. Includes an additional fix to handle a bad 401 response from vSphere
CIS session where the response does not include a WWW-authenticate
header (which is required according to HTTP spec). This causes the JDK
HttpClient to blow up with IOException before even handling the response
body. The quick fix is to treat the IOException as a status 401, which
will cause all invoking logic to retry once (and only once), after
attempting to reauthenticate.

2. Fixes a error were subnet was statically coded to /24 previously.