This is part-2 of my change. In this change, 'agent' is using 'awsutil' to read the node's config based on the cluster-id and node-id. It also uses 'docker-java' library for starting the ethrpc and concord-core.
This in my mind is boot-straping code. We should think a bit more about splititng out this code
as when we will do patch/upgrade, this code needs to be re-used to stop/start the containers.

This change you can consider as basic get something going but not the best organized.

Testing done:
We saw that 7-node deployment where agent starts the concord-node/ethrpc.

I have unit-tests for awsutil as well as for docker-client.