VB-745: do not force time publishing if tx published recently
The previous VB-745 commit introduced a thread to publish time
regularly. This commit prevents that thread from publishing when a
transaction already published time recently. This reduces the number
of "empty" blocks produced.

The mechanism is simply an atomic integer that is updated every time a
command including the time is sent to concord. The "pusher" thread
still wakes up every N milliseconds, but if that recorded time is less
than N milliseconds ago, there is no new time push needed, so the
thread just goes back to sleep.

The thread always sleeps for another N milliseconds, regardless of how
recently time was pushed. This means that we may not push time for up
to 2*period. We may want to change this in the future, but for now it
is a nice way to avoid extra time spent doing math and trying to sleep
for very short amounts of time.

One small bug is fixed in Helen API test, which was introduced in the
previous VB-745 change.

Testing is currently just a validation that no regressions have been
introduced. Inspecting test logs after a run shows fewer
zero-transaction blocks, which is an indication that the mechanism is
working. Further tests are on the horizon.