VB-577: Move tenancy from consoritum to organization
Main changes are to HelenUserDeatails and JwtTokenProvider.
Permitted chains is now computed from the current org.  The blockchains visible to consortiums this org is a member of is used now.

Refactored AuthHelper to be getOrganization, instead of getConsortium, and this is the reason for the number of files touched.

Updated some tests to look for this.  Also manually tested by creating another user to verify the blockchains are still visible.