Concord: Move kvb_storage to ethereum
This change moves the ethereum specific kvb_storage implementation to the
ethereum directory. This is in line with the current code organization to keep
everything Eth, DAML, or HLF specific in separate directories. Due to the move,
the blockchain directory is empty for now and hence the `concord_blockchain`
library disappeared. In turn, this change introduces a new library
`concord_ethereum` which packs all EVM/Ethereum details. Although this change
is mostly renaming, I had to adjust some includes and while doing so
encountered CRLF line endings which are addressed in the next commit.

FYI, here is a high-level overview of how data is supposed to move through
Concord: https://confluence.eng.vmware.com/display/BLOC/Concord+Data-flow

Further, the `concord/src/blockchain` directory will be filled when splitting
the existing `kvblockchain` libraray into a `consensus` and `blockchain` part.