[VB-873] Implementation of HLF with independent modules
This patch is to decouple the API services and KVBStorage instance of
HLF from the ETH implementation base on the discussion in MR!159.

Right now the implementation of HLF should be totally independent of
the Ethereum and DAML stack, and only one of them will be working in
the same time.

The following are the main works in this patch:
1. Implement API service for HLF in "src/hlf/concord_hlf_grpc_services".
2. Implement KVBStorage class in "src/hlf/concord_hlf_kvb_storage".
3. Update Dockerfile with go dependency.
4. Update "tools/conc_hlf_request.cpp" with gRPC client.
5. Update feature flag in the configuration system.