Concord: Adjust LF vs CRLF
In the Unix world and in most of our code base we use line feed (LF, \n) line
endings as opposed to the Windows world which uses carriage return line feed
(CRLF, \r\n). This change converts the line endings of those files that were
touched in the previous commit.