Concord: Blockchain -> concord::consensus
This change renames the `Blockchain::` namespace into `concord::consensus::` in
order to fit with the code structure. In a follow-up change every file under
`consensus/kvb/*` will move one level up and thereafter we split the logic into
a `concord::blockchain::` part and maybe a utils or common as well.