[VB-913] Fix enum type for orchestration resource events
This patch fixes an issue with respect to RESOURCE_EVENTS having
incorrect serialized int32 value.

There are 2 factors that contributed to the bug:
1. Protobuf IDL utilizes "reserved" keyword to reserve ranges in enum.
This was not done while skipping unused enum values in
provision_service.proto earlier.

2. kotlinx.serialization has an outstanding bug that does not allow enum
values to be skipped. This can be worked around. But it does not seem
worth the effort to invest at the moment.

This patch reorganizes the enum ordinal values for event type (in an
interface breaking fashion) and reslots RESOURCE_EVENT into sequential
ordering - while attempting to reorder the event types in a way that is
more "grouped" for currently available resource event types. i.e.
certian sequential event type ordinal values are somewhat "related".

The interface breaking change is permissible currently because the
component artifacts have not yet been versioned or released. And the
only downstream depenednecy besides tests is Helen, which is currently
only has source-level build dependency (i.e. API bindings are rebuilt
from source).