VB-920: transaction RLP should not include gas used
Gas-used was introduced to the RLP encoding of a transaction by commit
b02aefe4. This brown the ExtendedRPCTests eth_sendRawTransaction test,
but we didn't notice because it only warned about mismatched hashes,
instead of failing.

This change removes gas-used from the RLP encoding, and changes the
test to fail on mismatched hashes, so we don't break this again.

Gas-used is not part of the RLP encoding, because it's determined by
the miner that runs it, and not by the submitter. But, the submitter
must have signed the RLP encoding, so it must be the same both before
and after execution.

It's possible that this has been causing trouble with truffle and
other tools, because the transactions they submit would not have the
same hashes.