VB-885: initialize boolean value in detectLocalNode
Leaving hasDetectedNode uninitialized means that every once in a
while, it starts with the value `true`. That causes the node to fail
to start:

```
terminate called after throwing an instance of 'concord::config::ConfigurationResourceNotFoundException'
what(): Cannot determine which node configuration file is for: found private values for multiple nodes.
```

Valgrind finds this quickly:

```
blockchain/concord/build$ valgrind ./src/concord -c test/resources/concord2.config
==91592== Memcheck, a memory error detector
==91592== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
==91592== Using Valgrind-3.13.0 and LibVEX; rerun with -h for copyright info
==91592== Command: ./src/concord -c test/resources/concord2.config
==91592==
2019-05-23 16:14:53 91592 91592 DEBUG              Library:  94 | Successfully initialized RELIC library
==91592== Conditional jump or move depends on uninitialised value(s)
==91592==    at 0x4F40FE: concord::config::detectLocalNode(concord::config::ConcordConfiguration&) (configuration_manager.cpp:3786)
==91592==    by 0x4D2C0C: concord::config::loadNodeConfiguration(concord::config::ConcordConfiguration&, concord::config::YAMLConfigurationInput&) (configuration_manager.cpp:3695)
==91592==    by 0x4C7746: initialize_config(concord::config::ConcordConfiguration&, int, char**) (configuration_manager.cpp:66)
==91592==    by 0x48D264: main (main.cpp:416)
==91592==
```

In fact, the MemoryLeak Jenkins run has been reporting this for a
while. This spot in the code accounts for 3 of 11 of the instances of
this warning in the latest run. We should investigate the other 8
instances in the future, and see if we can clear them up so that we
can add this message as a failure trigger in the future.