Concord: mv consensus/kvb/* ..
In this change removes the `kvb/` directory by moving all files one level up.
With this change, some files (client, replica) are already in the correct
directory with proper names and namespaces due to the previous clean-up change.
In the next step we introduce the `concord_consensus` library which replaces
`kvblockchain` and finally thereafter split this library into a
`concord_storage` part.