VB-930: suppress uninitialized variable warnings in boost queue
Valgrind is complaining that boost::queue is using one of its
variables without initializing it first. This is a known issue that is
documented here: https://github.com/boostorg/lockfree/issues/48. It
has been determined to not actually cause problems, and also fixed,
but the fix hasn't yet been included in a release.

This change creates a valgrind suppression file for concord, and uses
it during the MemoryLeak test run, to prevent this unnecessary
warning.