VB-747: Default blockchain is not reading certs correctly
The default blockchain certs are actually stored in files.  When I refactored this code, I did not take that into account.
This change fixes that, and modifies the test to look for that.

Also this fixes VB-841, making it possible to start helen without the default blockchain.

Needed to fix MemberList to not read the cert from file.

Changed some logging to not include certs.

Reverting accidental change to copyright