Concord: Introduce `concord_storage`
This change moves storage related parts (mostly database interfaces) from
`consensus` into `storage` and creates a new library. At the moment, both
libraries are build the exact same way due to they need each other to build
which is a result of the split. However, in order to clearly separate, e.g. the
RocksDB dependecy, we need to refactor the code which should be a separate
change. Due to the split, this change introduces two new files to separate
between interface and implementation. Another minor change in here is the
renaiming from `release` to `releaseClient` & `releaseReplica` due to their
counter-part create functions.
In a follow-up change we should also split `blockchain_interfaces.h` due to it
defines both, consensus and storage interfaces.

* `blockchain_interfaces.h` should be split into a consensus and storage part as well