Fix Hermes's reference to provisioning service
There is a cosmetic change of the provisioning service's API stub name
from ProvisionService to ProvisioningService.

Background:
The original name as declared in protobuf IDL was ProvisionService.
During the container image and buildall.sh refactoring, it was decided
that ProvisionService is not grammatically correct, so the artifact name
was declared "persephone-provisioning" as opposed to
"persephone-provision".

The patch that introduces API level refactoring for persephone made the
source code name-change to better align with how the artifacts are
named.

This introduces a ripple effect for downstream consumers of Persephone
API, namely Helen and Hermes.