Concord: Make BlockchainDBAdapter manage IDBClient
The `createReplica` function takes a BlockchainDBAdapter object, creates a new
BlockchainDBAdapter, and makes sure that the database is initialized in order
to set a member and a member of a friend class.
In this change we untangle the dependecies, remove `createReplica` but still
enforce that the database is initialized. The `ReplicaImp` constructor takes
a `BlockchainDBAdapter` which takes an `IDBClient` and calls `init()`. To keep
it intuitive, we call `close()` when `BlockchainDBAdapter` gets destroyed.
Bonus: Fix memleaks.