[VB-888] Add ktlint to Persephone build
This patch adds ktlint, which is an anti-bikeshedding code formatter and
stylechecker for Kotlin code.

The default setting (ktlint by default runs configuration-file-less)
attempts to capture the official Kotlin language style guide + the
official Google Android style guide. For this project, which additional
style guides to follow is not very relevant nor applicable as the the
combination of the two above guides offer very reasonable and flexible
guidelines in producing readable Kotlin code.