[VB-904] Add provisioning_client script
This patch adds a provisioning client script to serve as an example/tool
to utilize for provisioning a Concord cluster.

This patch also modifies both metadata_client and provisioning_client to
allow both TLS-based and non-TLS-based communication with the server,
depending on the server's setting:

- If server expects TLS, client must also turn on TLS by supplying the
trusted certificate as command line argument.
- If server does not expect TLS, client can omit the trusted certificate
command line argument, and an insecure channel will be opened to the
server instead.

Lastly, the patch makes a minor logging enhancement to log orchestrator
initialization errors (errors are typically due to mis-configuration in
config.json).