[VB-991] Adopt ByteString protobuf serialization
This patch bumps up the internal protobuf library dependency to take in
the ByteString changes.

Specifically, the implementation of ProtoBuf encoder/decoder is now
based on a forked implementation of kotlinx.serialization's official
implementation, due to the fact that official implementation is not yet
capable of serialization of scalar bytes type for protobuf.