[VB-947] Introduce typed and untyped Value
This patch introduces TypedValue and UntypedValue as subtypes of Value
in the persistence KV layer. The main purpose is to create a base /
default implementation to be utilized by KV layer implementation. The
UntypedValue implementation is based off of the recently checked-in
protobuf ByteString implementation to properly represent an immutable
sequence of bytes (binary content).