WIP: DAML: Put PoC into CI/CD pipeline
Previous to this change, I uploaded a `daml-ledgerapi` image to artifactory
which is referenced in this change. The DAML PoC is special because it lives in
a separate repo and going forward we would like to keep it that way due to
collaboration with DA. This change updates the `docker/.env` file and makes a
special case in `make-prebuilt-env.sh`. The DAML test already exists and just
needs to be added to the CI/CD pipeline which is the purpose of this change.