[VB-991] Make protobuf serializer and ByteString more open
This patch further "opens" up ByteString to allow inheritance and
override of properties.

This patch also "opens" up protobuf so that the Encoder (ProtoBufWriter)
and Decoder (ProtoBufReader) are no longer marked as "internal" - so
that other downstream project modules can reference these classes to
build custom serializers that have special handling when dealing with
byte array and protobuf format.