Remove unused private variables in HLF classes
Clang 6 refuses to finish compiling because of unused variables in the
new HLF classes. This change removes those variables.

The removal of bft_sequence_num_ from kvb_storage needs to be
revisited, because it is actually an error that that value is not
stored. See https://jira.eng.vmware.com/browse/VB-1011.