Adding Spring Session to Helen
Spring Sessions gives us more control over the session manager. We will need it at some point anyway to have distributed session.

In the meanwhile, disabling a feature call Session Fixation, that changes the value of the session cookie on every call.  This is a problem when the browser fires multiple calls to REST endpoints at the same time.