VB-822: internal time service API
This change adds fields to TimeRequest (nee TimeUpdate) and
TimeResponse to allow directly reading the state of the time
contract. A commandline tool (conc_time) is added to allow use of this
message. If the time service is enabled, reading the time from a
containerized concord looks like this:

```
$ docker exec -it docker_concord1_1 ./conc_time
Message Prepared: time_request {
  return_summary: true
}

Connected
Message Sent (4 bytes)
Received response: time_response {
  summary: 1560359687625
}

The current time is: 1560359687625
```

If the time service is disabled, an error message is returned instead:

```
$ docker exec -it docker_concord1_1 ./conc_time
Message Prepared: time_request {
  return_summary: true
}

Connected
Message Sent (4 bytes)
Received response: error_response {
  description: "Time service is disabled."
}

Error Response: Time service is disabled.
```

A small fix to ExtendedRPCTests is also included, to ignore time-only
Ethereum blocks in the middle of a query range.

This API enables further testing of the time service, which will be
done in a separate JIRA task.