VB-1038: Convert Helen to use either log4j or slf4j, but not both
Helen currently had a mix of log4j and slf4j, but slf4j was not set up.  Consequently we were missing a lot of logging.

This MR converts all the slf4j to log4j.  The addition of ThreadContext to log4j covered the main reason to use slf4j (the MDC feature).