Move daml-on-concord into mono-repo
After meetings and discussion we decided to keep the mono-repo approach and
hence the standalone daml services are part of the mono-repo now. The new
directory contains two docker files which build the services. One is the
validator/execution engine and the other is the ledger api service. Building
both services has been integrated into `buildall.sh`.
The ledger api service in docker-compose-daml.yml is a hardcoded image in
artifactory which is the implementation from the first PoC. Going forward this
will be replaced with the one checked-in with this change.