[VB-947] Package test-jar for persephone's persistence lib
This patch adds additional packaging instruction to persephone
persistence's library to also package the module's test-jar.

Normally, this is not a universally accepted practice as packaging a
test-jar may introduce inadvertent dependencies to downstram on an
upstream module's tests.

In this case though, this is the *intended* behavior as persistence lib
defines a conformance test-driver that tests the basic SPI contract of
the persistence interface for any plug-able backend storage driver.