[VB-947] Force persistence KV layer's close() to be synchronous
This patch enforces a synchronous close() on persistence lib's key
value store layer. The original contract does not have a return value
(mimicking Closeable's contract). But this implies that the caller's
expectation is that by the time the call returns, all resources are
freed/released.

Failure to honor that contract on the implementation side, may cause
resource utilization races, particularly if the caller of close() then
releases additional resources that the KV implementation may still
depend upon (if in a 1/2 shutdown state).