[VB-947] Implementation of IP-allocation service
This patch implements an IP-allocator (light-weight IPAM) to control
private IP allocations within an orchestration site.

The service is exposed as a gRPC-based service, but is not intended to
be co-deployed with the management stack (Helen / Persephone). Instead,
it should be managed as a separate deployment that is meant to control
infra resources (IP).