Move all Helen API tests to PyTest.
Part of the larger change of creating a common PyTest base class
is to make the Helen API tests fit into that pattern.  This change
is a part of that; it moves the remaining Helen API tests to the
PyTest file.  (I'm trying to keep code reviews tolerable.)

Tests related to users were removed, as CSP handles users.