Remove metadata-service-client from codebase
Persephone metadata-service-client was added before there was a stable
python binding. The module was added mainly as a demonstration of what
was then the first gRPC based service client.

The module is not utilized by any CI/CD, which was the initial intended
use-case (called by Jenkins). Now with a stable python client binding
and an example metadata-server-client in persephone/resources/scripts/,
there is no longer any need to keep 2 client implementations around for
CI/CD to eventually integrate with.