Implement OrchestrationSiteService
This patch primarily implements OrchestrationSiteService, which is a
listing service to return orchestration site ID and other metadata
labels about each site back to the call (without returning sensitive
information like site credentials and backend infra URLs).

As part of the patch, provisioning server has been modified to host both
provisioning service and orchestration site service. This is ideal
because both services need orchestration site information. The server is
bootstrap/injected with the information, which allows it to do
provisioning operations as well as return the read-only view information
of the orchestration sites, depending on which service endpoint is
called.

Finally, there are some refactoring changes done, including moving
Dagger 2 related code into Kotlin instead of Java. Some of the verbosity
boilerplating in server bootstrap and injection code is reduced as a
result of this change. ProvisioningService implementation remains
untouched in this patch.