Add hardening improvements to concord agent's startup sequence
This patch makes various changes to how the concord agent's expectation
on the operating environment.

Specifically:
1. If the hosting VM already has the cluster configuration, do not reach
out to download the config again. This is useful in air-gapped on-prem
scenarios where the VMs connectivity may shift, depending on firewall
policies. Whereas at initial deployment time, public connectivity may be
available, subsequent start-ups may not have that connectivity
available.

2. If the hosting VM already has the container images locally, tolerate
the failure to re-fetch images with the same repository and label. This
is to facilitate similar requirements in air-gapped environments (see
1.)