Allow concord agent container to always restart automatically
This patch changes the initial launch parameter of concord agent such
that it will always restart unless it is manually stopped.

The implication is that even if the VM is shutdown, crashed and
restarted, or interrupted (at the VM level) during normal operations, as
long as the VM reboots, the Docker daemon will restart the agent
container as long as it (dockerd) is able to itself be started on boot
up.