Allow local TCP access to Concord node's Docker engine
This patch opens Docker API access at tcp://127.0.0.1:2375. The
implication is that any call that can reach the loopback interface on
port 2375 can interact with the Docker engine (start/stop containers).

Before service live, there will be another patch to open on port 2376
instead and engage in TLS authenticated communication only. For now, to
assist further development effort, the non-TLS port is used.

Security implication: Given that it is ONLY the local loopback
interface, there is NO remote access allowed via iptables rules. But any
logged in user does have effect root access to the system through
Docker. Right now, that implication is irrelevant because there are no
other users on the system except root. So if root user is logged in, the
implication is exactly the same.