VB-1045: Start/Stop API
Implmented controller to handle the start stop
Modified TaskController to accept a list of taskids to query

Made some changes in the way the TaskService works, so that it's easier to test

Added unit tests simulating GRPC callbacks.

Major refactoring of the api.yaml file to correct a huge number of errors in it.
Added API specifications for all the new REST calls, plus a couple that weren't documented

Used OpenApi generator to create the swagger.json file in the UI.  We need to make this part of the build step.

FleetService => FleetManagementService