Update to Spring Boot 2.1.6.RELEASE
Several reasons for this update

This is the version of Spring Boot that officially supports java 11
This supports an updated version of Mockito
Assorted security fixes

The Mockito version is important because it handles mocking final
classes.  This is needed to write unit tests that involve Persephone, since all of its stubs are final.