VB-833: teach conc_time to sign time updates
Since time updates now require signatures, we need the tool that the
tests use to send them to support signing.

Pass `-c <concord.config>` to `conc_time` to have it use the
`time_source_id` and signing key of the node for that config.

Also adds to the log message about missing values, to cover the
signature field.