VB-1172: Basic Auth fails occasionally with 401
The issue was a conflict trying to update the user's last login time.  This caused an exception, that lead to the authorization failing.

Fixed this problem in two ways: Change put to merge, and make any update of last login non-fatal

Also added chaching to basic auth
Renamed VmbcBasicAuthProver to VmbcBasicAuthProvider (oops)