VB-1173: Do not create TimeSigner if time service is not enabled.
TimeSigner insists that time_service_id is set. If time service is not
enabled, we don't otherwise need a time_service_id ... nor do we need
a TimeSigner.

This insistence was causing nodes deployed with configuration service
to fail to start, because they have time service off, and no
time_service_id defined.