Configurable concord-bft message size
With this change the message size, page size, and the maximum amount of
reserved pages can be specified in Concord's config file. It depends on !446
which updates concord-bft.
The idea is that the user can experiment with those values but wihtout any
guarantees. Chances are, if you don't know what you're doing changing the
values yourself will make concord fail at startup. However, those parameters
are not required and we fall back to defaults provided by concord-bft.