New DAML test
Our docker setup now relys on the ledger api test tool which is a separate
(manually uploaded) images that contains the java file. The jar itself contains
*.dar files which are required by at least one ledger api service at startup
time. When building the ledger api service, we pull the test tool and extract
the *.dar files so that it can be used for testing. This is only temporary
until a proper admin api is available to separately upload those packages.