VB-1059: refactor time service usage in prep for DAML&HLF (replica)
This is the first step of the refactor of Time Service for use with
DAML and HLF. In this change, a new ConcordCommandsHandler class is
introduced, and EthKvbCommandsHandler is changed to derive from it.

The ConcordCommandsHandler class takes over:

 1. Decoding the request buffer into a ConcordRequest.
 2. Handling the TimeRequest portion of the request.
 3. Encoding the ConcordResponse into the response buffer.

This is done like so:

 * ConcordCommandsHandler exposes the `execute` callback that is
   registered with ReplicaImp.

 * EthKvbCommandsHandler overrides a virtual Execute function that is
   called by ConcordCommandsHandler::execute.

 * ConcordCommandsHandler exposes an `addBlock` function, so that it
   and its subclasses implement the IBlocksAppender interface. In this
   function, generic data like the time contract state can be added to
   the updates being written in a block.

 * EthKvbCommands uses `this` as the IBlocksAppender that it passes to
   the EthKvbStorage constructor, so that when it writes blocks, those
   commands pass through ConcordCommandsHandler::addBlock, and get
   time updates added to them.

It is expected that HlfKvbCommandsHandler and
daml::KVBCCommandsHandler can inherit from ConcordCommandsHandler in
the same way.

Most of this change is moving the time contract code from
EthKvbCommandsHandler to ConcordCommandsHandler, and then reworking
EthKvbCommandsHandler::execute* into a new Execute function. This
provided the opportunity to remove the vestigial split between execute
and execute_read_only, that was required by early
concord-bft. Apologies for the sprawl of `const` additions to
otherwise unaffected functions.

We had talked about passing "just time" to the Execute function. This
change passes a pointer to a time contract instead, partly because
it's easier to deal with time service being either enabled or disabled
that way, and partly because it means we don't have to spend time
reading/summarizing the contract state if the execution never needs
it. This means we were unable to break the dependence of the ethereum
module on the time module.

ConcordCommandsHandler is in src/consensus, mostly because it's a
common place that most of our modules need to depend on
anyway. Because of its dependency on the time contract, it means the
consensus module depends on the time module.

The dependency of the time and time test modules on ethereum *has*
been broken!

The change to the time tests looks a little weird because, after
switching from using EthKvbStorage to the raw
ILocalKeyValueStorageReadOnly interface, problems developed with the
key comparator:

 1. It's odd that this ever worked, because the key comparator takes
    two arguments, not three.

 2. The comparator expects a block ID in the key, which is added by
    the DB interface used by the replica, which is not part of the
    mock storage in this test.

The main TODO spread throughout this change is about VB-1011. Now that
we have a common place to store data that every execution environment
should maintain, we can move the BFT sequence number storage out of
EthKvbStorage and into ConcordCommandsHandler. This commit was already
getting large, so this is punted to a future commit.