VB-1011: move BFT sequence number storage to ConcordCommandsHandler
Part of the BFT-persistence story is to store the BFT sequence number
in the KVB block, along with the smart contract data written in that
block. This has been implemented in our Ethereum storage for some
time, but has been missing from HLF and DAML.

This change moves the storage of the sequence number out of
EthKvbStorage and into a new generic ConcordMetadataStorage that is
used by the recently-added ConcordCommandsHandler. Ethereum, HLF, and
DAML, all get sequence number storage by default.

Two subtle changes:

 * I moved the sequence number storage prefix from 0x07 to 0x21. This
   moves it out of the Ethereum range (0x0?) and into the generic
   range (0x2?), next to the time service storage. This should be
   safe, because we are not persisting data across deployments yet.

 * "Sequence Number == 0" is no longer written to the Ethereum genesis
   block. This should be safe, because GetBlockMetadata returns 0 if
   no entry is found.