TimeService: Use google::protobuf::Timestamp
So far the time service dealt with `uint64_t` but the unit of those numbers
wasn't encoded anywhere and the code using the time service assumed
milliseconds. Needless to say, this is error prone if we add more time sources
and/or change our time pusher. Therefore, this change changes the type from
uint64_t to google::protobuf::Timestamp in order to be agnostic to the time
unit. In addition, time signing and ordering is now done via nano seconds.
The benefit is that, time sources can provide time in any unit they want as
long as it can be encoded via a Timestamp. Similar, consumers inside concord
can choose the time unit which make sense in their context.