Remove docker file changes to pull in config-gen
we donot need config gen tool anymore here. However, the code related to
that is not removed but marked as dead code since it was decided to keep
it hanging around until we have a closure on whether we could have the
entire s3 workflow out of picture. We could alternatelyuse config
service even if s3 workflow is maintained, but that might be less
performant. So keeping the code around.