Adds CLI option to conc_genconfig for outputting a JSON principal ID map.
This commit adds a command line option, specifically --report-principal-locations, to conc_genconfig, which causes it to output a JSON file with a mapping of (node index)->(array of principal IDs in no particular order) describing which Concord-BFT principals are on which node in the cluster that configuration is generated for. This feature is intended primarily for use by deployment. This commit also includes a unit test of the function implementing the generation of this JSON given a ConcordConfiguration that validates the JSON output agrees with the contents of the configuration.