Increase ledger api test tool timeout
We see frequent timeouts for the first test of the ledger api test tool but
there is no evidence that there is an implementation issue. When we
experienced the timeout locally then it was due to that transaction where is
still in progress. This change increase the timeout by factor 2.