VB-1097: Allow client requests to time out
We have been passing "infinity" as the timeout value to
SimpleClient->sendRequest. This means that our clients never give up
on requests that they send to the network.

This change introduces a configuration parameter
`bft_client_timeout_ms` that allows us to specify an alternative
timeout, in milliseconds, so that clients will give up after some
time. The default for this config is `0`, which causes the current
"infinity" value to be used.

Concord unit tests have been changed to pass a five second timeout
value, as a simple smoke test. Integration (hermes) tests are left
with infinite timeouts for the moment.