[VB-1104] Modify provisioning service to consume agent changes
This patch completes the set of commits that integrates concord agent
with configuration service. (Note: This patch does NOT actually
integrate provisioning service with configuration service. It merely
consumes the necessary changes that allows agent to integrate with
configuration service. The agent portion which is already checked in
is currently disabled and pending actual integration of configuration
service with provisioning service, in order to be able to be turned on).

This patch also use the newly created CA-signed IPAM service
certificate, and modifies Dockerfile accordingly to install the CA
certificate that signed the IPAM service certificate.